+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Remove usage of session and session rehydration from AdminAuthenticator plug.
 
 ## 0.14.0
 - Add 401 unauthorized response to Webhook plug